api_kursus

+++ src/pendaftaran/controller.js
@@ -1,16 +1,15 @@
-const {
-  tambahPendaftaran,
-  hapusPendaftaran,
-  ubahPendaftaran,
-  lihatPendaftaran,
-  cariIdPendaftaran,
-} = require("./service.js");
+const { tambahPendaftaran, hapusPendaftaran, ubahPendaftaran, lihatPendaftaran, cariIdPendaftaran } = require("./service.js");
 
 const createPendaftaran = async (req, res) => {
   try {
-    const { tanggal_daftar,  siswa_id, kursusId ,status_pembayaran} = req.body;
+    const { tanggal_daftar, kursus_id, status_pembayaran } = req.body;
 
-    const body = { tanggal_daftar,  siswa_id, kursusId ,status_pembayaran };
+    const body = {
+      tanggal_daftar,
+      siswa_id: req.user.id,
+      kursus_id,
+      status_pembayaran,
+    };
 
     const data = await tambahPendaftaran(body);
     return res.status(200).json({ message: "Data berhasil ditambahkan", data });
@@ -19,10 +18,14 @@ const createPendaftaran = async (req, res) => {
   }
 };
 
-const getPendafataran = async (req, res) => {
+const getAllPendaftaran = async (req, res) => {
   try {
     const data = await lihatPendaftaran();
-    return res.status(201).json({ message: "Data pendaftaran", data });
+
+    return res.json({
+      message: "Data Pendaftar",
+      data,
+    });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -31,8 +34,16 @@ const getPendafataran = async (req, res) => {
 const deletePendaftaran = async (req, res) => {
   try {
     const id = req.params.id;
-    const data = await hapusPendaftaran(id);
-    return res.status(200).json({ message: "Data berhasil dihapus" }, null);
+    const data = await cariIdPendaftaran(id);
+    if (!data) {
+      return res.status(404).json({ message: "Maaf, Data tidak ditemukan" });
+    }
+
+    if (req.user.role !== "admin" && data.siswa_id !== req.user.id) {
+      return res.status(403).json({ message: "Maaf, Anda tidak punya akses" });
+    }
+    await hapusPendaftaran(id);
+    return res.status(200).json({ message: "Data berhasil dihapus" });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -42,9 +53,14 @@ const getAllById = async (req, res) => {
   try {
     const id = req.params.id;
     const data = await cariIdPendaftaran(id);
-    return res
-      .status(201)
-      .json({ message: "Data pendaftaran berdasarkan id", data });
+    if (!data) {
+      return res.status(404).json({ message: "Data tidak ditemukan" });
+    }
+
+    if (req.user.role !== "admin" && data.siswa_id !== req.user.id) {
+      return res.status(403).json({ message: "Akses ditolak" });
+    }
+    return res.status(201).json({ message: "Data pendaftaran berdasarkan id", data });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -53,8 +69,21 @@ const getAllById = async (req, res) => {
 const updatePendaftaran = async (req, res) => {
   try {
     const id = req.params.id;
-    const { tanggal_daftar,  siswa_id, kursusId ,status_pembayaran } = req.body;
-    const body = { tanggal_daftar,  siswa_id, kursusId ,status_pembayaran };
+    const dataLama = await cariIdPendaftaran(id);
+    if (!dataLama) {
+      return res.status(404).json({ message: "Data tidak ditemukan" });
+    }
+
+    if (req.user.role !== "admin" && dataLama.siswa_id !== req.user.id) {
+      return res.status(403).json({ message: "Maaf, Anda tidak punya akses " });
+    }
+    const { tanggal_daftar, kursus_id, status_pembayaran } = req.body;
+    const body = {
+      tanggal_daftar,
+      siswa_id: dataLama.siswa_id,
+      kursus_id,
+      status_pembayaran,
+    };
 
     const data = await ubahPendaftaran(id, body);
     return res.status(200).json({ message: "Data berhasil diubah", data });
@@ -64,7 +93,7 @@ const updatePendaftaran = async (req, res) => {
 };
 module.exports = {
   createPendaftaran,
-  getPendafataran,
+  getAllPendaftaran,
   deletePendaftaran,
   updatePendaftaran,
   getAllById,
